# NodeClient/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Film Manager",
    "description": "REST APIs for the interaction with the Film Manager service.",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3001/"
    }
  ],
...
  "paths": {
...
    "/api/films/public/{filmId}/images": {
      "post": {
        "tags": [
          "api/films/public/{filmId}/images"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Associate a new image to a public film",
        "description": "The image sent in request body is associated to the public film characterized by the ID specified in the path. Only the film owner can associate an image to the film.",
        "operationId": "addImage",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "image to assign to the film",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "image": {
                  "contentType": "image/png, image/jpg, image/gif"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      },
      "get": {
        "tags": [
          "api/films/public/{filmId}/images"
        ],
        "summary": "Retrieve the list of all the images that have been associated to a public film",
        "description": "All the images that been associated to the public film with ID filmId are retrieved as a json-encoded array of image data structures. Only the film owner or a film reviewer can retrieve the images.",
        "operationId": "getImages",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film whose images must be retrieved",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Images"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      }
    },
    "/api/films/public/{filmId}/images/{imageId}": {
      "get": {
        "tags": [
          "api/films/public/{filmId}/images/{imageId}"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Retrieve an image associated to a public film",
        "description": "The image with ID imageId, associated to the film with ID filmId, is retrieved. In particular, by specifying the desired content type via the Accept header, the user can decide whether to retrieve the image data structure (json content type), which does not contain the image file, or the image file itself, in one of the supported image content types (image/png, image/jpg, and image/gif). Only the film owner or a film reviewer can perform this operation.",
        "operationId": "getSingleImage",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "imageId",
            "in": "path",
            "description": "ID of the image",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/jpg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/gif": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "414": {
            "description": "Unsupported Media Type",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": [
          "api/films/public/{filmId}/images/{imageId}"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Delete an image associated to a public film",
        "description": "The image to be deleted is the one with ID imageId, associated to the film with ID filmId. Only the film owner can delete the image.",
        "operationId": "deleteSingleImage",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "imageId",
            "in": "path",
            "description": "ID of the image",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Image": {
        "type": "object",
        "properties": {
          "$schema": {
            "type": "string",
            "description": "schema for Image data structure"
          },
          "id": {
            "type": "integer",
            "description": "unique identifier for the image"
          },
          "name": {
            "type": "string",
            "description": "name without extension of the image"
          },
          "self": {
            "type": "string",
            "format": "uri-reference",
            "description": "self link"
          }
        }
      },
      "Images": {
        "description": "schema for Images data structures",
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Image"
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "description": "JWT to identify the authenticated user.",
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    }
  }
}